refactor: :recycle: Move messageContent default state into a hook
